requirements.txt
torch~=1.6.0
opencv-python~=4.4.0.42
numpy~=1.18.5
tqdm~=4.46.1
inpainting~=1.0.0
pillow~=7.1.2
torchvision~=0.7.0
pandas~=1.0.5
setuptools~=46.2.0
scipy~=1.3.2
scikit-image~=0.17.2
flowiz~=2.3.1
typing~=3.7.4.3